intermittent commit
