+++ tests/integration/test_auditlogs.py
@@ -3,6 +3,9 @@ from datetime import timedelta
 
 import pytest
 from tests.integration import run_command
+from tests.integration.util import DockerDaemon
+from tests.integration.util import SyslogServer
+
 
 SEARCH_COMMAND = "code42 audit-logs search"
 BASE_COMMAND = "{} -b".format(SEARCH_COMMAND)
@@ -10,6 +13,37 @@ begin_date = datetime.utcnow() - timedelta(days=-10)
 begin_date_str = begin_date.strftime("%Y-%m-%d %H:%M:%S")
 end_date = datetime.utcnow() - timedelta(days=10)
 end_date_str = end_date.strftime("%Y-%m-%d %H:%M:%S")
+
+
+begin_date = datetime.utcnow() - timedelta(days=-10)
+begin_date_str = begin_date.strftime("%Y-%m-%d %H:%M:%S")
+
+
+@pytest.fixture
+def data_transfer():
+    with DockerDaemon():
+        print("docker daemon started")
+        with SyslogServer():
+            print("syslog server started")
+            yield run_command
+
+
+@pytest.mark.integration
+@pytest.mark.parametrize(
+    "command",
+    [
+        (
+            "code42 audit-logs send-to localhost -p TCP -b '{}'".format(
+                begin_date_str
+            )
+        )
+    ],
+)
+def test_auditlogs_send_to(data_transfer, command):
+    exit_status, response = data_transfer(command)
+    print(exit_status)
+    print(response)
+    assert exit_status == 0
 
 
 @pytest.mark.integration
@@ -45,3 +79,6 @@ end_date_str = end_date.strftime("%Y-%m-%d %H:%M:%S")
 def test_auditlogs_search_command_returns_success_return_code(command):
     return_code, response = run_command(command)
     assert return_code == 0
+
+
+
